io: Switch fair_queue_ticket on 512 block size
Currently tickets are 2-dimentional: (weight, size). The size is accounted
in bytes and is multiplied by a value, that depends on write-to-read
throughput relation times "magic" constant of 128. As the result summing
up tickets for e.g. 256k requests may overflow the 32-bit counter quite
fast easily.

Nowadays it's not a big problem, the overflow only happens if the disk
capacity is set to unlimited, but with IO groups approach too fast overflow
will cause problems with fairness.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>