fair_queue: Do group-based resource limiting
The idea of the scheduling resembles the "sliding window" algo. Each
group has two rovers -- tail and head. The distance between them is the
available capacity.

When a queue needs to dispatch it moves the tail ahead, thus decreasing
the remaining capacity. When the request completes it increments the
head.

When grabbing the capacity tail is NOT checked against the head, so it
can overrun one. This means, that the queue is over-dispatched and the
gap between head and tail (reversed one) denodes the capacity demand.

The next dispatch is possible when the capacity is back, but just waiting
for the head to run ahead of the tail will work like dumb spinlock --
queues will lose races and will continue spinning on rovers for too long.
Instead, the queue waits for the head to overrun some local value, which
is the tail + needed capacity taken once. Then the queue continues its
dispatching, so they line-up at the order at which they shift the tail
counter, similar to how the ticket-lock makes spin-locks more fair.

Sometimes queue overflows the capacity (and goes pending state) with
some request, but before the waiting finishes it tries to dispatch some
other request with different ticket. In this case the pending is just
rearmed from the scratch.

Another thing to remember is that just polling the head rover for the
needed value may cause big delays if reactor goes the interrupt mode. To
wake it up the queue arms the highres timer at some time point which is
estimated from the fair-ticket value and the "pace" value, which, in
turn, is the reciprocal of the disk speed (bytes/sec or ops/sec). The
highres timer poller goes before io poller, so we're likely to dispatch
the request in the same task quota that expires the timer.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>