ioq: Prepare the ioq poller for wakeup
Currently the io queue poller doesn't care when the reactor enter
interrupt mode. It needs to be polled only if either a new request
comes (which happens in non interrupt more) or when a previous
request completes (which wakes reactor from this mode).

With io groups the situation is different. The io-queue may decide
that it cannot dispatch right now, because the shared capacity it
exchausted, but that it should (try to) dispatch some time later.
This "some time later" is not connected with submitting new or
completing old requests on local shard, so we need a way to wake
up the reactor from sleep for pure io-queue needs.

This option is already possible with lowres-timers poller, so
borrow the wakeup timer from there.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>