repeat: Simplify loop
The allocation of repeater this patch removes was correct, but
unnecessary.

When it was done we had no pending future, so we could convert an
allocation failure into exceptional future. That was not obvious, and
I first thought this patch was fixing a bug.

The allocation was redundant, since set_callback works with both
available and unavailable futures.

Removing it also allows some followup simplifications.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200820000251.246505-2-espindola@scylladb.com>