posix: add one more setsockopt overload
Current overloads assumed that the passed data is in form of a string,
while sometimes (e.g. SO_TIMESTAMPING) it points to an integer,
so running strlen() on it in order to get its length is not applicable.
Also, in most cases it's probably possible to elide strlen() entirely,
since the length is already known.
Message-Id: <88e3043dbcb6ce94c97e02ca8ad4862749969573.1597937123.git.sarna@scylladb.com>