testing: Stop boost from installing its own signal handler
Normally seastar uses a signal handler that prints a backtrace and
then unregisters itself so that the program crashes and produces a core
dump.

Unfortunately, SEASTAR_TEST_CASE uses the boost test framework, which
also installs signal handler. Doing that disables the backtrace
printed by seastar and also disables the crash, so we don't get a core
file.

This changes registers a dummy handler before calling into boost. With
that a failing test in dev or release mode also prints a backtrace and
produces a core.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200820170559.178518-1-espindola@scylladb.com>