Add a Seastar_DEBUG_ALLOCATIONS build option
With it seastar will memset allocations before returning the
pointer. The reason I want this is because a new future still in the
future state will have most of whatever value was using that address
before.

Those stale values make it a lot harder to track pointers in a core
dump.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200821223740.395683-1-espindola@scylladb.com>