future: Don't put a std::tuple in future_state
This changes future and related types to not store a std::tuple in
future_state. Given how odd void is, this patch adds
internal::monostate, which is the same as std::monostate without the
need to include the variant header.

Since these changes are user visible, they are behind a new API level.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200828011030.1912895-5-espindola@scylladb.com>