http: tests: change misleadingly named server startup functions
In multiple places in the httpd tests there are "writer" functions
which are in fact simply creating a http request handler and starting
a server that uses it. This patch changes these names to "server_startup"
or simplifies it even further by predefining the handler that is used
repeatedly and running the former function code without wrapping in
a lambda.

Signed-off-by: Wojciech Mitros <wojciech.mitros@scylladb.com>