core/reactor: complete_timers(): restore previous scheduling group
complete_timers() currently resets the current scheduling group before
returning. This is needed because we set the current scheduling group
for each timer, so continuations created by them inherit the scheduling
groups of the timer, so after all the timers are processed we need to
reset it. We currently reset it to the default group assuming
`complete_timers()` is only called from system context. However this is
not always true, in particular `run_tasks()` can end up calling it
indirectly via `reactor::reset_preemption_monitor()`. This is bad
because `run_tasks()` sets up the current scheduling group to be the
same as that of the task queue it is currently running. This results in
the current scheduling group changing in the middle of a `run_tasks()`
call, resulting in some tasks observing an incorrect current scheduling
group.

Refs: scyllasb/scylla#7117
Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20200903124900.1063831-1-bdenes@scylladb.com>