rpc: Return actual chosen compressor in server reponse - not all avail
Fixes #790

When using compression in RPC, the multi-compressor factory is used.
This has a feature set of the compressors put into it.

Negotiation goes:

* Client sets COMPRESSION in initial request to a comma separated list of
  the compressors it can handle.
* Server resolves this, which typically means the first compressor in list
  is picked from its multicompressor. Server sets its compressor object from this.
* Server responds by setting COMPRESSION in response to its supported list.
* Client picks the first/highest prio compressor it can support from this list and
  sets compressor object.

Server compressor response should be based on actual compressor chosen, not the
available set.

Note that multi_algo_compressor_factory today deals with this
via the "is_server" parameter, but this still makes assumptions
on the other side choosing the first matching algo etc. I.e.
the negotiation is still shaky, and for no valid reason as
the server already has set his compressor, and there is no
id for that in subsequent frames.

This patch simply adds a "name" attribute to compressors,
and changes the server response to be this instead of
supported compressor set. Also ensures we throw an error
on client side if server picked a compressor we cannot
handle.

Message-Id: <20200908110654.28655-1-calle@scylladb.com>