future: fix uninitialized constexpr in is_tuple_effectively_trivially_move_constructible_and_destructible
The template variable is_tuple_effectively_trivially_move_constructible_and_destructible
is only defined for tuples, and is left undefined for non-tuples. gcc is happy with this
(the variable is only instantiated for tuples), but clang appears to check more eagerly.
I believe that clang is probably correct in rejecting this uninitialized constexpr
variable, so prevent the problem by indirecting via a helper class template, which delays
the check.
Message-Id: <20200912170823.1065745-1-avi@scylladb.com>