alien: be compatible use API Level 5
in API level 5, internal::monostate is used to represent the state
of the return value of a function which returns "void", and tuple
is not used for passing the return value anymore. so in this
change, alien is updated to follow this convention and accept
futurize<>::value_type instead of futurize<>::tuple_type, otherwise
the alien appication fails to compile if API Level 5 is enabled, in that
case, instance of type T or internal::monostate is passed to
return_type_of::set() as the second parameter, and compile chokes
at seeing this.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20200913143729.3525628-2-tchaikov@gmail.com>