net: tls: Added feature to register callback for TLS verification
The callback provides only very basic information to the handler,
but one advantage is that its API is independent of the current
TLS backed (currently gnutls, might change in the future).

The registration method is part of tls::certificate_credentials,
so that both clients and servers can make use of this feature.

I'd like to thank Piotr, Calle and Avi for their suggestions for
improvement.

Signed-off-by: Niek Bouman <niek.git@gmail.com>
Message-Id: <20200914121102.11941-1-niek.git@gmail.com>