shared_future, shared_promise: document constructors' noexcept-ness
The compiler propagates the noexcept property to constructors
and operators defined as `= default`, but it is hard for the user
of the library to understand that from reading the header file.

Add comments in the code documenting the noexcept property
rather than explicitly specifying it (e.g. with "noexcept = default;")
and staticly assert that in futures_test.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>