doc: fix hyper link to tutorial.html
tutorial.html is created at ${CMAKE_CURRENT_BINARY_DIR}/tutorial.html
where $CMAKE_CURRENT_BINARY_DIR is the build directory of seastar/doc.
while per doc/Doxyfile.in, the output directory specified by
`HTML_OUTPUT` is "html". and the working directory of CMake target of
`doc_api` is also ${CMAKE_CURRENT_BINARY_DIR}. so by default,
the generated hyperlink in index.html is pointing to current directory
of "doc/html/index.html", i.e. "doc/html/tutorial.html". which does not
exist.

so in this change, we need to point the browser to ../tutorial.html
which is the output file of md2html.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20200924162239.87378-3-tchaikov@gmail.com>