doc: update split script to use xml.etree.ElementTree
the htmlsplit.py failed to split the HTML generated with pandoc 2.9.2.1.
because of the change in the HTML schema. in pandoc 2.9.2.1, the
generated TOC looks like

<nav id="TOC" role="doc-toc">

instead of

<div id="TOC">

so the existing script fails to split the chapters from the table of
conetents, and sends all the generated document to split/index.html.
we'd better off matching the "id" attribute of the element.

instead of using regex to match the output, use xml.etree.ElementTree to
parse and modify the XML tree in hope that it'd be simpler to adapt to
future changes in the output schema.

in this change, instead of iterating the generated HTML document line by
line, this script

1. collect /head and /body/header element, and use it as a template
2. add the toc to the template for the new index.html, then
3. put all /body/h1 elements into separate html files under the
   output directory named as <chap-num>.html

this script does not change the format of the output, it just parses and
dumps the document.

this script was tested using pandoc 2.7.3 and 2.9.2.1

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20200929125338.30707-2-tchaikov@gmail.com>