core/shared_ptr: Expose use count through {lw_}enable_shared_from_this
The owning pointer object already exposes it.

In some cases we also need to access it from the object being shared,
which doesn't have access to the owning pointer.

For example, some caches could use lw_shared_ptr to implement sharing,
but when object is not shared then it is owned by something else,
e.g. an LRU. use_count() in this case could be used to determine the
owner.
Message-Id: <1601653923-29465-1-git-send-email-tgrabiec@scylladb.com>