memory: align allocations to std::max_align_t
The allocator currently ensures that allocations are aligned to
the greatest power of two that divides the size. For example, size
8 allocations are aligned to 8, size 16 allocations are aligned
to 16, and size 48 allocations are also aligned to 16 (48 = 16*3).
This is natural since types are padded to their alignment, and
alignment must be a power of two.

However, clang makes use of the standard's specification that
malloc returns memory aligned to std::max_align_t (16 for x86_64),
even for allocations that aren't a multiple of that size. It does
so by using aligned vector mov instruction (movaps) even though
there is no sub-object in that allocation that could possibly be
aligned to 16.

Since this behavior is legal, indulge it and align anything above
std::max_align_t to std::max_align_t. Smaller allocations are left
unaligned, it's unlikely that clang will require alignment larger
than the object size. This reduces memory waste.
Message-Id: <20201004081711.2522915-1-avi@scylladb.com>