smp service group: add optional group name
Currently when one of the group's semaphores throws
a timeout exception, the exception message contains
the groups id which is not very indicative. This commit
adds an optional name parameter to the smp group
configuration parameters. If this name is set, it will
be indicated in the smp group's semaphores timeout exception
messages.
The reason for this parameter to be optional is so old code
using smp groups or validation systems relying on the exceotion
message format will not break.

Tests: unit tests (dev mode)