semaphore: use functor as wait list expiry_handler
This allows us to instantiate a single ExceptionFactory -
the semaphore's base-class, rather than two of them,
one for the semaphore and one for the expiry_handler.

With that. constructing the wait list becomes noexcept
(based on std::function being nothrow move constructible)

Note that theoretically, exception_factory::timeout() may throw.
To simplify error handling in the expiry handler function,
catch this exception and turn the error into a generic, non-throwing one
and with that mark the returned function noexcept.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>