logger: allow applying a rate-limit to log messages
Sometimes one wants to dump a detailed diagnostics report on an error to
help diagnosing the underlying problem. Just logging these with >= info
is problematic because these reports can be big and the errors can come
in bursts (or simply happen too often), completely flooding the logs.
Currently the standard workaround for this is to log these with debug
level and have the user set the log level to debug when they are
interested in the diagnostics. This is also problematic, this doesn't
help with diagnosing problems that already happened and reproducing
might be hard or impossible (and will be prone to log flooding just the
same).

This patch presents a solution to this problem: a flexible way to
rate-limit log messages. The rate-limiter can be shared with multiple
messages, across multiple loggers and it ensures that in the given
interval at least one of the messages sharing the rate-limiter will be
logged. This might seem overly-restrictive, but for the diagnostics dump
use-case it was developed for it is perfectly reasonable as there will
be typically just one message.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20201030060524.431137-1-bdenes@scylladb.com>