resource: Collect orphan PUs to be assigned later
Sometimes a processing unit (PU in hwloc terms) may be attached
to a NUMA node with no memory on board (e.g. the respective
ram banks are empty). In this case newer (2.2.0+) hwloc does NOT
report any NODE and seastar hits the respective assertion. Older
hwloc reported empty node.

To fix this we'll need to distribute node-less PUS (called
orphans in this patch) among existing nodes and for this it's
needed to collect them all when mapping PUs to nodes.

Next patch will make the distribution itself.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>