resource: Distribute orphan PUs among nodes
When a PU is reported by hwloc not to have any local NUMA node
it should better be assigned to some "random" node, rather than
crash. To do this assignment group all the orphan PUs, then
scatter the groups between nodes evenly.

The grouping can be done in many ways, the goal here is to have
even distribution. On the other hand it seems worth keeping the
PUs sharing the Lx cache in the same node. So the grouping
logic is -- try to group PUs by L3, L2 and L1 caches sequentally,
until it's seen that we cover all the available nodes.

If even L1 is not enough -- then scatter PUs one-by-one.

When this situation happens, warn user in logs, that the hardware
is not configured/assembled in optimal way.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>