memory: fix malloc_usable_size() during early initialization
malloc_usable_size() checks whether it is running in a reactor
thread in order to see if it needs to dispatch to the libc
malloc_usable_size() or our own. But in one case this fails.

The case is when a static constructor performs memory allocation.
at this point, the reactor and memory system have not been
initialized yet, so is_reactor_thread == false. The original
libc memory functions have not been looked up yet, since these
operations also happen in a static constructor, which can
be scheduled after the one triggering the problem.

Fix by using is_seastar_memory() instead. This is similar to
what realloc() does, for the same reasons.
Message-Id: <20201105164841.3858222-1-avi@scylladb.com>