util: introduce critical_alloc_section.hh
This file contains `scoped_critical_alloc_section` which allows marking
scopes containing critical allocations as such. This is a generalization
of the current `disable_failure_guard`, and in fact it replaces it
deprecating the former. The reason we are generalizing
`disable_failure_guard` is that we want to attach more mechanics to
critical allocations, then just disabling allocation failure injection
for them.