build: clean up after failed -fconcepts detection
The cmake code sets CMAKE_REQUIRED_FLAGS to "-fconcepts" in order
to see if that flag is available. But if not (as is expected with
clang), it leaves CMAKE_REQUIRED_FLAGS, and so every following test
becomes poisoned, since it will fail on the unsupported flag.

This break the membarrier test which follows. As a result membarrier()
will not be used even if it is available. This is more or less okay
on x86 which has a fallback (even there, the fallback is pretty bad
on large machines), but on aarch64 there is no fallback, so Seastar
programs spin even when they have nothing to do.

Fix this by clearing the flag after the test. In order to change
as little as possible, if the test succeeded, we set it again.
Message-Id: <20201119185453.1729601-1-avi@scylladb.com>