scheduler: Count total waiting time for task queues
The wait time is the time the queue was spent without
tasks in it. It starts when the queue goes off the active
list and ends when the activate() is called on it.

The new _ts member on task queue is named like that
because it will be reused by the next patch.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>