scheduler: Count total starvation time for task queues
The starvation time is the time during which the queue
needs a cpu, but doesn't get one for any reason. The
startvation starts when the scheduler picks some other
queue to run tasks from or when the activate() is called
on the queue. Starvation ends when the queue is selected
for execition.

This time is useful when finding out the reasons for
increased latencies. Comparing waittime vs starvetime
point to whether the workload is io- or cpu-bound.

The _ts member is used to note deactivation time, but
since queue cannot be in both -- wating and starving --
states at the same time this field can be reused to
note the starvation start time.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>