append_challenged_posix_file_impl::read_dma: allow iovec to cross _logical_size
Currently append_challenged_posix_file_impl::read_dma
trims the iov vector if it goes beyond _logical_size
causing short reads if the file length is unaligned.

We have to allow the iovec covering EOF to extend
beyond it up to an offset aligned on disk_read_dma_alignment.

Add a unit test that reproduces this case and passes
with the fix.

Test: unit(dev)
Signed-off-by: Benny Halevy <bhalevy@scylladb.com>
Message-Id: <20201201062849.3995029-1-bhalevy@scylladb.com>