util: noncopyable_function: avoid gcc uninitialized error in move constructor
In direct mode, we use bytewise copy (allowed since we check for
is_trivially_move_constructible). This copies holes and padding in the
object, and gcc notices this and complains.

Fix by silencing the warning for that code section.
to store the user function. The move constructor copies all of those 32
bytes, even the unused ones. This is because copying a small fixed size
buffer is faster than a variable sized buffer, and because we don't store
the used size of the buffer anywhere.
Message-Id: <20201213121912.1559226-1-avi@scylladb.com>

Fixes #845.