shared_ptr: define debug_shared_ptr_counter_type constructor as noexcept
In debug mode, share_ptr uses debug_shared_ptr_counter_type (instead of
a plain integer) as its counter type. Its constructor, however, was
not defined as noexcept. This propagates to enabled_shared_from this, and
thence to async_sharded_service. This causes async_sharded_service's
default constructor to be deleted, since it is required to be noexcept
and is defined with "= default". As a result, distributed_test's
async_service doesn't have a default constructor, and does not compile
under gcc 9.3.

Fix be marking debug_shared_ptr_counter_type's constructor as noexcept.
This is safe since std::this_thread::get_id() (used for a member) is
also noexcept.

Fixes #845.
Message-Id: <20201213124438.1568654-1-avi@scylladb.com>