io_queue: Turn fair-queue lambda into a callable
The fair queue code queues noncopyable_function's into its
proirity queues. While dispatching, it peeks the functions
from its internal queues and calls.

In case of IO queue the noncopyable_function is per-request
one which is constructed from a lambda, that captures the
io_request itself, a bunch of helper variables and whose
code is to call sink.submit().

Other than being not very convenient for cancellaion of a
request, this construction is memory-consuming, as the
capture in question doesn't fit the noncopyable_function's
inline size of 32 bytes and allocates the capture block
separately. Effectively it takes 40 + 48 = 98 bytes on
top of the internal::io_request to queue one. Changing
this part by queueing the captured data directly would
halve the per-request queue overhead.

First introduce a queued_io_request class that's inherited
from internal::io_request and adds all the helper data from
the mentioned capture block. Before queuing it directly,
construct a noncopyable_function out of it and queue as
before (to make this work -- declare the operator() for the
new guy).

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>