fair_queue: Turn fair-queue callable into an object
The second part of converting a fair-queue lambda into an
explicit object.

This time introduce a fair_queue_entry hook structure that's
queued into the fair_queue and carries a C-style callback
that's called upon entry dispatch. The caller is expected to
find out the container of the entry itself, likely with the
boost::intrusive get_parent_from_member() call.

The queued_io_request that was intruduced in the previous
patch is the first (and only) used of it.

This patch completes the request queue rework. As the result
it saves 40 bytes for each queued request, and makes it
possible to collect them into a "cancellation queue".

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>