io_queue: Cancellable queue
The nowadays queue of requests look like this:

queue -> request -> request -> ... -> request -> (nul)

and this list is implemented with the intrusive singly linked
list. To support the cancellation of individual requests the
requests must sit in yet another list, which would look like
a skip-list-like subset of this one. Like this

queue -> request -> request -> ... -> request -> request -> (nul)
                     ^  |              ^  |
handle --------------+  +----  ...  ---+  +--> (nul)

When "pressed" the handle should walk its list and somehow
invalidate the requests. For space-efficiency it would be nice
that this 2nd list is also a singly-linked list, but this
kind of list does not allow to withdraw elements from the
middle of it.

So the cancellable queue design is:

1. requests are queued in singly linked list on handle
2. when dispatched from the main queue, the request is
   removed from the handle's cancellation queue, because
   it can only be at its head
3. when cancelled, the request remains in the main queue,
   because it can be in the middle of it

Also, when dispatching the request, there's no cancellation
handle at hands, so requests should have back-reference on
their handles. But since dispatching only happens from the
head of the main queue, and thus -- from the head of the
relevant handle, a pointer on the handle should only sit
on the first request from the handle's list.

This makes the cancellation queue look like this:

handle  <---(backref)--+
 |                     |
 +- first --> request -+
 |
 `-  rest  --> request --> request -> ...

Where the "rest" is, again, a singly-linked-list.

When queued the request gets into the first pointer or, if
it's already taken, to the back of the rest.
When de-queued, the first request leaves the queue and gets
replaced with the front of the rest list.
The backrefer and next pointers are unioned on the request
for memory efficiency. Thus the cancellation queue adds a
single pointer over a queued request (given we've previously
saved 40 bytes from it, this is still great).

When destroyed, the cancellable queue cancels all the requests
from its list. There's no separate method for it, as the queue
is intenral io_queue.cc object, the API-level handle will come
a bit later.

When the request is cancelled it releases all it can right
at once. This includes resolving the promise on whose future
the IO-completion is waited for. Only the request itself stays
in the main queue until it gets dispatched in its order, then
it's freed completely.

The cancelled request resets its fair-queue ticket to zero not
to make the fair queue account this one into the capacity
engine. With this change the cancelled request are dispatched
as noop and just free the memory they occupy.

Both changes solve two goals:

- the IO buffer is released right at cancellation time
- the request presence affects neither the real disk IO
  nor the fair-queue counters that might affect other
  requests

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>