io_queue: Track requests in cancellable queue
The io_queue::queue_request() method now accepts a pointer to
the io_intent object. If not null, the request being submitted
it "attached" to the intent. If later the intent is destroyed
or .cancel()-ed, all the requests will complete right at once.

It's important to obey the sequence:

1. queue the request into main queue
2. no-throwing-ly queue the request into the cancellation queue

because it's impossible to undo the enqueue operation to any
of the queues.

To make it work, first the io_intent is asked for the exact
cancellation queue reference and after the potentially throwing
main queue enqueue succeeds, the request is pushed back to the
cancellable queue in noexcept manner.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>