memory: Use 2^n sizes when searching for preferred span size
Currently, a page-granular search is used for the preferred span
size for small pools: we search for a span size that has little
waste when chunked up into objects of the size handled by this
pool. However, returned spans are always power-of-two, so this
search is not making the right calculation.

This fix changes the search to only consider power-of-two sized
spans, since those are the only ones available. This does change
the preferred span size for many pools, especially larger ones.
When the preferred span size does change, it is always larger.

The span chosen is the smallest span with less than 5% waste,
i.e., internal fragmentation, or if no such span size exists
within the max span size of 32 pages, we choose the span size
with the lowest total waste.

Fixes: #836

Signed-off-by: Travis Downs <travis.downs@vectorized.io>
Message-Id: <20201215064529.559858-1-travis.downs@vectorized.io>