fair_queue: Change the dispatch condition
Canceled requests will sit in fair-queue until get dispatched.
Not to affect and/or get stuck in the capacity checks, the
entry's ticket will be reset to zero value for them and the
queue's "queued resources" counter will be changed according
to this.

But right now this counter is used to decide whether or not
to proceed with dispatching requests. With a queue full of
cancelled requests this will not happen until the next queued
request.

It's, probably, not critical, but it's still better to process
the queue earlier, so check for the number of queued requests
(not resources) instead. This counter will continue to operate
as before even for canceler requests.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>