coroutine: deprecate coroutine forwarding
The syntax

    co_return a_function_that_returns_a_future(p1, p2);

saves a task switch, but is a time bomb as any parameters (p1, p2)
are not held stable on the coroutine frame and instead must be copied
or moved. Since it's easy to invoke this accidentally, so deprecate it.

In addition, for the future<void> case, the syntax is not legal. Clang
and gcc 11+ reject it, but gcc 10 accepts it. Tighten the checks for the
compiler to avoid failures on gcc 11.

The unit test is removed.

There is no replacement yet, apart from the less-efficient
"co_return co_await ...".
Message-Id: <20201216125202.2829158-1-avi@scylladb.com>