fair_queue: Move dispatching callback into dispatch_requests
The fair queue is carrying the information how to handle its
entries on the entries themselves. Although this can be quite
convenient if we want to handle different entries differently,
this makes the entry 30% larger in size.

At the same time even with cancellable IO (next patches) there
will be only two options of handling an entry, so things get
simpler after removing the dispatching callback from the entry
and passing it into fair_queue::dispatch_requests.

The entry itself becomes one pointer thinner. Times the number
of requests in the queue gives good memory savings.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>