io_request: Remove kernel completion from it
The io_request carries the kernel_completion pointer inside
to be used by io_sink code -- the completion is copied onto
the aio cb. Before getting to the sink queue the request sits
in the io queue and this pointer is unused, but takes space.

Introduce the pending_io_request version of the io_request
that has the completion on board and remove the completion
from the io_request.

This saves some per-request memory and prepares the ground
for cancellation of pending requests (not in this set though).

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>