io: Bring the io_intent argument onto file level
Currently the io_queue::queue_request accepts an intent pointer
argument and defaults it to nullptr. Remove the defaults and
make all the callers provide intent (nullptr as well for now)
themselves.

These callers are tests and the only posix_file_impl methods.
All are internal seastar stuff, so the library API is not
changed.

Making the posix_file_impl itself accept the intent argument
requires a bit more trickery, and thus comes in the next patch.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>