io_queue: Limit tickets for oversized requests
The IO queue is not expected to effectively schedule requests
whose size is bigger than a certain limit. Usually 256K per request
is more than enough to satisfy both -- consume all disk throughput
and maintain reasonable latency for urgent requests.

Respectively, fair group is configured to pass through the limited
sum of requests. And if one shard sends a request which is alone
larger (in size) than the group limit it will never get dispatched.

Handle this by capping the fair queue ticket value assigned to a
request so that it doesn't eat more than half of the group limit.
This "half" would typically be several megabytes which is far above
the expected usage.

tests: io_tester with 32Mb requests

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20201230132752.16032-1-xemul@scylladb.com>