io_tester: Make file sizes configurable
By default io_tester runs all shards in all workloads on 1Gb file.
This is not enough, write workload does shows different results
depending on the file size and 1Gb is not "big enough to work".
It's tempting to have the ability to change the size of the data.

At the same time specifying the size of per-shard-per-workload
file is not very convenient, as the available disk space is most
likely a constant value. So when changing the the number of shards
for the test one would need to update the conf file for every run,
which is annoying.

Said that -- here's the configuration option  called "data_size"
which is the disk space assigned for the whole workload,  which
is split evenly between all shards inside it.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20210105061553.6481-1-xemul@scylladb.com>