file: preserve dma alignment values when transporting a file across shards
Seastar has a file_handle mechanism for reopening a file on another
shard. Unfortunately, it did not preserve the dma alignment values,
leading to over-aligned reads for files opened this way (in very
unlikely cases, it could lead to under-aligned writes, but this
requires unconventionally formatted filesystems on systems with a
large page size).

Fix by copying the alignment values to the file handle and using them
there.