abort-source: Improve the subscriptions management
There are two concerns about the abort_source subscriptions:

1. The subscription carries a pointer to the abort_source, but
   this pointer is not updated if the abort source moves. The
   pointer is unused, however, so it's not critical

2. Aftter the abort source is fired, the associated subscriptions
   continue resolving to true, as if they are armed. This is not
   critical either, as callers are expected to check this right
   after the subsribe() call

However, the 2nd item looks like a time bomb -- if the steps by
chance happen to be

  auto s = as.subscribe()
  as.request_abort()
  if (s) {
      ...
  }

the ... piece is better not to be called.

Both can be fixed and the subscription itself can be made thinner
if removing the back-pointer to the abort source completely. The
cast to bool is replaced with the check if the subscription is
linked.

tests: unit(dev)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20210111120854.7505-1-xemul@scylladb.com>