io_intent: Safe reference
The intent object is "encouraged" to get destroyed at any time, but
some file implementations will need to carry the reference on it
through preemption points.

Said that -- here's the safe reference on an intent. It doesn't
perform any additional allocations and throws the "cancelled"
exception on attempt to revive the original intent.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>