file: Extend the API with intents
Now when the file_impl is ready to accept io_intents it's
time to extend the top-level API with them. This time with
yet another argument with default "not-cancellable" value
for compatibility and those not willing to cancel the IO.

Together with the previous patch the intent is ignored, but
this is going to be fixed really soon.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>