tls: Make reloadable certificates more robust
Changes the file watch to wait for "close", i.e. someone
was _done_ with file (write or copy/move), as well as use
directory polling more for if files are deleted (which a copy
over will often look like).
We also keep better track of how well we've succeeded to reduce
the somewhat eager error reporting.
This means we should never try to report of failures because
an individual file was half-written. Note though that some files
like cert/key pairs will still be sensitive to races, since
only one of them will typically be changed at a time.
This will be addressed in next patch.