fair_queue: Preempted requests got re-queued too far
When a request is preempted, it's put at the group's queue
tail by setting the queue to wait for the group's head rover
to become larger than the group's current tail. In turn,
when the preempting request completes it forwards the group
head rover by _its_ size.

If the preempting request is smaller than the preempted one,
and there are more than one queue doing this, the resulting
head value may not become large enough to pass by the expected
thresholds and the whole group may get stuck.

The problem was introduced by 98366fd5 and it reveals itself
when queue is fed with over-sized requests and they all are
preempted with tiny ones.

Before the mentioned patch this didn't happen because preempted
request pushed the groups head rover forward as if they had
completed, thus letting the queues dispatch their requests (not
in order, but still).

The fix is in re-queueing preempted requests right after the
preempted one.

tests: unit(dev), dtest.resharding(dev)

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>
Message-Id: <20210126144325.14086-1-xemul@scylladb.com>