http: add reading/skipping all bytes in an input_stream
If an application knows that it will need all bytes from a buffer,
it can now use short_stream_reader methods:
the read_entire_stream() method saves all buffers from a stream to
a std::vector.
Some applications may know that they're going to receive small amounts
of data in a stream. For them it may be more convenient to work on a
single string instead of a vector. The read_entire_stream_contiguous()
method provides such a string.
An app (e.g. an http handler) may need to exhaust an entire input_stream
(e.g. to enable processing the next http request on the same connection).
This can now be achieved by using the skip_entire_stream() method.

Refs: scylladb/scylla#7213

Signed-off-by: Wojciech Mitros <wojciech.mitros@scylladb.com>