http: add ragel parsers for chunks used in messages with Transfer-Encoding: chunked
In preparation for chunked request handling, additional parsers are needed
because the chunk format is strictly specified by IETF. The parsers follow
instructions from the RFC 7230.

The body of a message with Transfer-Encoding: chunked is composed of zero
or more chunks, followed by the last chunk, trailer headers, and a CRLF token.
Each chunk starts with the chunk size, chunk extensions and CRLF - the
http_chunk_size_and_ext_parser matches them and provides methods that
allow to extract them: get_parsed_extensions() and get_size(). The
remainder of the chunk does not need to be parsed.
The trailer headers and the CRFL that's ending the message can be parsed
by the http_chunk_trailer_parser and accessed using the get_parsed_headers()
method.

Signed-off-by: Wojciech Mitros <wojciech.mitros@scylladb.com>