tls: reloadable_credentials_base: add_dir_watch: fix root dir detection
The existing implementation will spin forever if add_watch
throws on every directory since if breaks the loop with an
exception only when !dir.has_parent_path().  However,
if dir is an absolute path, beginning with '/', it will
always have parent_path (only if it's a relative path
traversing up the path will reach an empty parent_path).

Therefore, break from the loop with an exception either
if there no parent_path (in the relative path case), or
when we reached the root path, where dir.parent_path() == dir.

This might happen when fsnotifier::impl::create_watch
keeps throwing std::runtime_error("attempting to use closed notifier")
if not active.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>
Message-Id: <20210131111802.54345-1-bhalevy@scylladb.com>