reactor: optimize need_preempt() thread-local-storage access
Currently, need_preempt() uses an extern thread_local variable.
This forces the compiler to refer to it through the global offset
table (although, since we compile with -fvisibility-inlines-hidden,
it can be argues that this is a compiler bug).

Instead, change it to use a function-local static variable. This
makes it clear to the compiler that it cannot be overridden by
a dynamic library (such as with LD_PRELOAD), and the compiler
refers to the variable directly, instead of through the global
offsets table. Since need_preempt() is used in numerous places,
this can reduce the instruction count.
Message-Id: <20210206175717.952496-1-avi@scylladb.com>